Fix folders that include ignored files
If a folder includes a file that has been ignored, then it would almost
always be shown as "ignored", even if the folder also contains files
that have NOT been ignored.

An example file structure:

```
.
├── .gitignore
└── folder
    ├── ignored.txt
    └── not_ignored.txt
```

Where the `.gitignore` file includes the following:

```
folder/ignored.txt
```

This causes the whole `folder/` entry to be marked as "ignored" even
though `not_ignored.txt` is not ignored. This doesn't seem to make sense
to me and I think that a folder should be marked as "ignored" _only_ if
all of the files inside it are ignored.